sound2/requirements.txt
numpy
pandas
librosa
soundfile
torch
torchaudio
scikit-learn
tqdm
pyyaml
matplotlib
